+++ CacheServer/src/emem/common/rmi/CacheClient.java
@@ -9,7 +9,7 @@ import java.util.Map;
  * 其中hash是一套键值对集合, 键和值都是字符串类型. 可以操作整套的hash数据, 也可以对hash数据中单个的索引进行操作.
  * 缓存没有提供remove键的操作, 但对每个数据进行了过期设置, 例如给定过期设置expire=10, 则10秒后数据会从缓存中清除.
  * 每个数据操作接口会给出两种版本, 一种是不含过期参数, 另一种是含有的. 例如set方法:
- *      set(String key, String value) 和 set(String key, String value, long expire)
+ *      set(String key, String value) 和 set(String key, String value, int expire)
  * 两种. 另外, 针对普通HTTP协议的jar封装, 虽然获取具体的CacheClient实例的方法不同, 但是CacheClient所含的接口方法
  * 是一致的, 完全可以参照本接口的方法定义来对照出相关的方法定义.
  *
@@ -76,20 +76,20 @@ public interface CacheClient extends Remote {
     long hashSize(String key) throws RemoteException;
 
 
-    void set(String key, String value, long expire) throws RemoteException;
+    void set(String key, String value, int expire) throws RemoteException;
 
-    String get(String key, long expire) throws RemoteException;
+    String get(String key, int expire) throws RemoteException;
 
-    void hashSetAll(String key, Map<String, String> map, long expire) throws RemoteException;
+    void hashSetAll(String key, Map<String, String> map, int expire) throws RemoteException;
 
-    Map<String, String> hashGetAll(String key, long expire) throws RemoteException;
+    Map<String, String> hashGetAll(String key, int expire) throws RemoteException;
 
-    void hashSet(String key, String index, String value, long expire) throws RemoteException;
+    void hashSet(String key, String index, String value, int expire) throws RemoteException;
 
-    String hashGet(String key, String index, long expire) throws RemoteException;
+    String hashGet(String key, String index, int expire) throws RemoteException;
 
-    void hashRemove(String key, String index, long expire) throws RemoteException;
+    void hashRemove(String key, String index, int expire) throws RemoteException;
 
-    long hashSize(String key, long expire) throws RemoteException;
+    long hashSize(String key, int expire) throws RemoteException;
 
 }
